Extract Cineast REST calls, remove unwanted IDs from features
