+++ src/main/kotlin/ch/unibas/dmi/dbis/vrem/generation/GenerationConfig.kt
@@ -3,7 +3,7 @@ package ch.unibas.dmi.dbis.vrem.generation
 import kotlinx.serialization.Serializable
 
 enum class GenerationType(val cineastCategory: String = "", val tableName: String = "") {
-    SEMANTIC_SOM("som_semantic", "features_visualtextcoembedding"), // Only contains 1 feature.
+    SEMANTIC_SOM("som_semantic", "features_visualtextcoembedding"),
     VISUAL_SOM("som_visual", "features_AverageColor"),
     SEMANTIC_SIMILARITY("sim_semantic"),
     VISUAL_SIMILARITY("sim_visual"),
